+++ src/lib/project-tasks/index.ts
@@ -5,6 +5,7 @@ import * as bumpMinor from "./bump-minor";
 import * as bumpPatch from "./bump-patch";
 import * as lint from "./lint";
 import * as lintFix from "./lint-fix";
+import * as tsconfigJson from "./tsconfig-json";
 import * as tslintJson from "./tslint-json";
 
 export function registerAll(gulp: Gulp, project: Project) {
@@ -13,6 +14,9 @@ export function registerAll(gulp: Gulp, project: Project) {
   bumpPatch.registerTask(gulp, project);
   lint.registerTask(gulp, project);
   lintFix.registerTask(gulp, project);
+  if (project.typescript !== undefined && project.typescript.tsconfigJson !== undefined) {
+    tsconfigJson.registerTask(gulp, project);
+  }
   tslintJson.registerTask(gulp, project);
 }